+++ dns/src/qclass.rs
@@ -34,7 +34,9 @@ impl TryFrom<u16> for Qclass {
     type Error = u16;
 
     fn try_from(value: u16) -> Result<Self, Self::Error> {
-        match value {
+        // mdns responses will sometimes have the high bit set
+        // https://datatracker.ietf.org/doc/html/rfc6762#section-18.12
+        match value & 0b0111_1111 {
             x if x == (Self::IN as u16) => Ok(Self::IN),
             x if x == (Self::CS as u16) => Ok(Self::CS),
             x if x == (Self::CH as u16) => Ok(Self::CH),